backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
requests
google-genai